# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Mock FPS API",
    "description": "Mock server simulating the Form3 Faster Payments Service (FPS) API. Returns realistic JSON:API responses with async lifecycle status transitions, stand-in mode, and webhook notifications.",
    "version": "0.1.2",
    "contact": {
      "name": "hum3",
      "url": "https://codeberg.org/hum3/mock-fps"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
    "/v1/transaction/payments": {
      "post": {
        "tags": ["Payments"],
        "summary": "Create payment",
        "operationId": "createPayment",
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentDataEnvelope"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDataEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" }
        }
      },
      "get": {
        "tags": ["Payments"],
        "summary": "List payments",
        "operationId": "listPayments",
        "parameters": [
          {
            "name": "filter[submission.settlement_cycle]",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "filter[submission.settlement_date]",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "filter[admission.settlement_cycle]",
            "in": "query",
            "schema": { "type": "integer" }
          },
          {
            "name": "filter[admission.settlement_date]",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "page[size]",
            "in": "query",
            "schema": { "type": "integer", "default": 100 }
          },
          {
            "name": "page[number]",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "List of payments",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentListEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transaction/payments/{paymentID}": {
      "get": {
        "tags": ["Payments"],
        "summary": "Get payment",
        "operationId": "getPayment",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "responses": {
          "200": {
            "description": "Payment found",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentDataEnvelope"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/v1/transaction/payments/{paymentID}/returns": {
      "post": {
        "tags": ["Returns"],
        "summary": "Create return",
        "operationId": "createReturn",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnPaymentDataEnvelope"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Return created",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnPaymentDataEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" }
        }
      },
      "get": {
        "tags": ["Returns"],
        "summary": "List returns",
        "operationId": "listReturns",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "responses": {
          "200": {
            "description": "List of returns",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnPaymentListEnvelope"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/v1/transaction/payments/{paymentID}/recalls": {
      "post": {
        "tags": ["Recalls"],
        "summary": "Create recall",
        "operationId": "createRecall",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/RecallDataEnvelope"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Recall created",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallDataEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" }
        }
      },
      "get": {
        "tags": ["Recalls"],
        "summary": "List recalls",
        "operationId": "listRecalls",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "responses": {
          "200": {
            "description": "List of recalls",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallListEnvelope"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/v1/transaction/payments/{paymentID}/reversals": {
      "post": {
        "tags": ["Reversals"],
        "summary": "Create reversal",
        "operationId": "createReversal",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/vnd.api+json": {
              "schema": {
                "$ref": "#/components/schemas/ReversalDataEnvelope"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reversal created",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ReversalDataEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "415": { "$ref": "#/components/responses/UnsupportedMediaType" }
        }
      },
      "get": {
        "tags": ["Reversals"],
        "summary": "List reversals",
        "operationId": "listReversals",
        "parameters": [
          { "$ref": "#/components/parameters/paymentID" }
        ],
        "responses": {
          "200": {
            "description": "List of reversals",
            "content": {
              "application/vnd.api+json": {
                "schema": {
                  "$ref": "#/components/schemas/ReversalListEnvelope"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "paymentID": {
        "name": "paymentID",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/vnd.api+json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/vnd.api+json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "UnsupportedMediaType": {
        "description": "Content-Type must be application/vnd.api+json",
        "content": {
          "application/vnd.api+json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "Resource": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "id": { "type": "string", "format": "uuid" },
          "organisation_id": { "type": "string" },
          "version": { "type": "integer" },
          "created_on": { "type": "string", "format": "date-time" },
          "modified_on": { "type": "string", "format": "date-time" }
        },
        "required": ["type", "id"]
      },
      "AccountParty": {
        "type": "object",
        "properties": {
          "account_name": { "type": "string" },
          "account_number": { "type": "string" },
          "sort_code": { "type": "string" }
        }
      },
      "ChargesInformation": {
        "type": "object",
        "properties": {
          "bearer_code": { "type": "string" }
        }
      },
      "FxInfo": {
        "type": "object",
        "properties": {
          "contract_reference": { "type": "string" },
          "exchange_rate": { "type": "string" },
          "original_amount": { "type": "string" },
          "original_currency": { "type": "string" }
        }
      },
      "RelationshipData": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "id": { "type": "string", "format": "uuid" }
        },
        "required": ["type", "id"]
      },
      "Relationship": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RelationshipData" }
          }
        }
      },
      "PaymentRelationships": {
        "type": "object",
        "properties": {
          "payment_submissions": { "$ref": "#/components/schemas/Relationship" },
          "payment_admissions": { "$ref": "#/components/schemas/Relationship" },
          "payment_returns": { "$ref": "#/components/schemas/Relationship" },
          "payment_recalls": { "$ref": "#/components/schemas/Relationship" },
          "payment_reversals": { "$ref": "#/components/schemas/Relationship" }
        }
      },
      "PaymentAttributes": {
        "type": "object",
        "properties": {
          "amount": { "type": "string", "example": "100.00" },
          "currency": { "type": "string", "example": "GBP" },
          "end_to_end_reference": { "type": "string" },
          "numeric_reference": { "type": "string" },
          "payment_scheme": { "type": "string", "default": "FPS" },
          "payment_type": { "type": "string" },
          "processing_date": { "type": "string", "format": "date" },
          "reference": { "type": "string" },
          "scheme_payment_sub_type": { "type": "string" },
          "scheme_payment_type": { "type": "string", "default": "ImmediatePayment" },
          "beneficiary_party": { "$ref": "#/components/schemas/AccountParty" },
          "debtor_party": { "$ref": "#/components/schemas/AccountParty" },
          "charges_information": { "$ref": "#/components/schemas/ChargesInformation" },
          "fx": { "$ref": "#/components/schemas/FxInfo" }
        },
        "required": ["amount", "currency"]
      },
      "PaymentResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/PaymentAttributes" },
              "relationships": { "$ref": "#/components/schemas/PaymentRelationships" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "PaymentDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/PaymentResource" }
        },
        "required": ["data"]
      },
      "PaymentListEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PaymentResource" }
          }
        },
        "required": ["data"]
      },
...
      "ReturnPaymentAttributes": {
        "type": "object",
        "properties": {
          "amount": { "type": "string", "example": "100.00" },
          "currency": { "type": "string", "example": "GBP" },
          "return_code": { "type": "string" },
          "return_reason": { "type": "string" },
          "scheme_status": { "type": "string" }
        },
        "required": ["amount", "currency"]
      },
      "ReturnPaymentResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/ReturnPaymentAttributes" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "ReturnPaymentDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/ReturnPaymentResource" }
        },
        "required": ["data"]
      },
      "ReturnPaymentListEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReturnPaymentResource" }
          }
        },
        "required": ["data"]
      },
...
      "RecallAttributes": {
        "type": "object",
        "properties": {
          "amount": { "type": "string" },
          "currency": { "type": "string" },
          "recall_reason": { "type": "string" },
          "recall_type": { "type": "string" },
          "status": { "type": "string" }
        }
      },
      "RecallResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/RecallAttributes" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "RecallDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/RecallResource" }
        },
        "required": ["data"]
      },
      "RecallListEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/RecallResource" }
          }
        },
        "required": ["data"]
      },
...
      "ReversalAttributes": {
        "type": "object",
        "properties": {
          "amount": { "type": "string" },
          "currency": { "type": "string" },
          "reversal_reason": { "type": "string" },
          "status": { "type": "string" }
        }
      },
      "ReversalResource": {
        "allOf": [
          { "$ref": "#/components/schemas/Resource" },
          {
            "type": "object",
            "properties": {
              "attributes": { "$ref": "#/components/schemas/ReversalAttributes" }
            },
            "required": ["attributes"]
          }
        ]
      },
      "ReversalDataEnvelope": {
        "type": "object",
        "properties": {
          "data": { "$ref": "#/components/schemas/ReversalResource" }
        },
        "required": ["data"]
      },
      "ReversalListEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ReversalResource" }
          }
        },
        "required": ["data"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "status": { "type": "string" },
                "title": { "type": "string" },
                "detail": { "type": "string" }
              },
              "required": ["status", "title"]
            }
          }
        },
        "required": ["errors"]
      },
...
    }
  }
}